Cache: Create a caching class
In case the GIFView gets the GIF image via URL, and not as an
InputStream, the GIFView needs to take over the responsibility of
caching the downloaded file. If if doesn't, it will re-download the
image file on each rotation and other configuration change. This
behavior is extremely undesirable. The new class GIFCache acts like a
proxy between the internet and the View, taking care of downloading the
image and providing a cached version of it. When a cached version of the
image is found, it will be delivered, otherwise it will be downloaded
first, and then the cached version will be delivered.