cache: Adding sub-directory in cache-area
This change introduces a sub-directory for the app's cache area in which
to store all cached files. The sub-directory is fixed, however, the
implementing app can easily retrieve the sub-directory as a File object
by calling getCacheSubDir(). This allows the app to easily delete all
the cached contents, as the directory will be recreating when GIFCache
detects that it is missing. GIFs will then of course be downloaded
again.