cache: getCacheSubDir() is now public / static
For easier access to getCacheSubDir(), the class is now public and the
method is static. It requires the Application-context in case the
cacheSubDir has not yet been initialized, as it will be initialized then
on the fly.

Furthermore, a section has been added to README.md that explains the
caching, how it works and how the implementing app can actually get rid
of the cached files again.